requirement.txt
streamlit
python-dotenv
langchain
langchain-huggingface
langgraph
pydantic
huggingface-hub